+++ frontend/src/event/JoinButton.js
@@ -1,0 +1,9 @@
+import { Button, Row } from "react-bootstrap";
+
+export default function JoinButton() {
+  return (
+    <Row className="justify-content-xl-center mx-5">
+      <Button>Sign me up</Button>
+    </Row>
+  );
+}
